chore: move lego to its own package
